Re-factored classes to inherit from a single base class
